TIP-860: Move AttributeOption related api tests
