+++ tests/back/Pim/Structure/EndToEnd/AttributeOption/ExternalApi/PartialUpdateAttributeOptionEndToEnd.php
@@ -1,12 +1,12 @@
 <?php
 
-namespace Akeneo\Tool\Bundle\ApiBundle\tests\integration\Controller\AttributeOption;
+namespace AkeneoTest\Pim\Structure\EndToEnd\AttributeOption\ExternalApi;
 
 use Akeneo\Test\Integration\Configuration;
 use Akeneo\Tool\Bundle\ApiBundle\tests\integration\ApiTestCase;
 use Symfony\Component\HttpFoundation\Response;
 
-class PartialUpdateAttributeOptionIntegration extends ApiTestCase
+class PartialUpdateAttributeOptionEndToEnd extends ApiTestCase
 {
     public function testHttpHeadersInResponseWhenAnAttributeOptionIsUpdated()
     {
